rest/flask/pyproject.toml
[tool.poetry]
name = "rest"
version = "0.0.0"
description = "REST API Template"
authors = ["ChangZhuo Chen (陳昌倬) <czchen@czchen.org>"]
license = "MIT"

[tool.poetry.dependencies]
connexion = {version = "^2.14.2", extras = ["swagger-ui"]}
flask-migrate = "^4.0.4"
flask-sqlalchemy = "^3.0.5"
hydra-core = "^1.3.2"
openapi-spec-validator = "^0.6.0"
opentelemetry-instrumentation-flask = "^0.40b0"
prance = "^23.6.21.0"
python = "^3.11"
python-json-logger = "^2.0.7"


[tool.poetry.group.dev.dependencies]
testcontainers = {version = "^3.7.1", extras = ["postgresql"]}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
